refactor[#523] : entities/profile queryKey constants 에서 api 로 이동
