+++ src/entities/profile/api/index.ts
@@ -1,1 +1,2 @@
 export * from "./getProfile";
+export * from "./queryKey";
